+++ src/Model/Catalog/Layer/Filter/Item.php
@@ -47,6 +47,8 @@ class Item extends MagentoItem
      * @param Filter $filter
      * @param AttributeType $attributeType
      * @param Url $url
+     * @noinspection MagicMethodsValidityInspection
+     * @noinspection PhpMissingParentConstructorInspection
      */
     public function __construct(Filter $filter, AttributeType $attributeType, Url $url)
     {
@@ -181,5 +183,13 @@ class Item extends MagentoItem
     public function isActive()
     {
         return $this->getAttribute()->getIsSelected();
+    }
+
+    /**
+     * @return string
+     */
+    public function getCssId()
+    {
+        return spl_object_hash($this);
     }
 }